Refactor Datasets and Notebooks context to enhance state management and streamline component interactions
