+++ DashAI/front/src/components/custom/contexts/DatasetsAndNotebooksContext.jsx
@@ -74,6 +74,8 @@ export const DatasetsAndNotebooksProvider = ({ children }) => {
     removeNotebooksByDatasetId,
     selectedOption,
     setSelectedOption,
+    step,
+    setStep,
   };
 
   return (
