+++ DashAI/front/src/hooks/datasets/useDatasets.js
@@ -152,21 +152,16 @@ export function useDatasets({ t }) {
     datasets,
     selectedDatasetId,
     enrichDatasetsWithInfo,
-
     createDataset,
-
     fetchDatasets,
     fetchFreshDatasets,
     selectDataset,
     clearSelectedDataset,
-
     deleteDatasetLocal,
     deleteDatasetRemote,
-
     editDataset,
     addDatasetOptimistically,
     startDatasetPolling,
-
     removeDatasetById,
     replaceDatasets,
   };
